Move LanguagePicker component to dedicated file
